rename button_list to ButtonList to be consistent with other components
