+++ apps/src/code-studio/pd/workshop_survey/VariableFormGroup.jsx
@@ -1,6 +1,6 @@
 import React, {PropTypes} from 'react';
 
-import {ButtonList} from '../form_components/button_list';
+import {ButtonList} from '../form_components/ButtonList';
 import FieldGroup from '../form_components/FieldGroup';
 
 import {
